+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Auto-focus the two step verification field
 * Do not escape the LoginGuard method title (allows for title formatting, e.g. with the backup codes method) 
 
+**Bug fixes**
+
+* The emergency backup codes could be reused
+
 # 1.2.0
 
 **Other changes**